chore: use our private Nix cache for Docker image pushes
This is a public repo, but for technical reasons, our CI system only
pushes PR build artifacts to our private Nix cache. When the ghcr.io
Docker image push job runs in GitHub Actions, on merges to `main`, we
want to be sure that job doesn't rebuild anything from scratch, so we
need to give this job access to our private Nix cache.

Note that a future commit will push Nix build artifacts to our public
Nix cache on merges to `main`, but it won't fix this particular issue,
because that public cache push job and the Docker image push job will
run simultaneously.

Signed-off-by: Drew Hess <src@drewhess.com>