feat: read list of input derivations from `stdin`
Haskell isn't a particularly good replacement for a shell, so let's
use the best tool for the job. We now read the list of input
derivations to build from `stdin`, and rely on an external
tool(s) (currently still just `nix-instantiate`) to generate that
list. This has a number of advantages over the previous approach:

1. There are numerous tools for generating lists of input derivations,
so now we're not tied to `nix-instantiate` and `nix-build`.

2. Most of these tools generate JSON, so we can use the wonderful `jq`
to massage their output before passing the simple list of derivations
to `nix-buildkite`.

(For the time being, we still hard-wire `nix-instantiate` in the
plugin hook, but that will be easy to fix in a subsequent commit.)

Note that we don't bother to use a fancy streaming library for reading
`stdin`. It wouldn't do us any good, because we need the full list of
input derivations (all in memory at the same time) in order to build
the dependency graph.

Signed-off-by: Drew Hess <src@drewhess.com>