feat: support a `nix build` mode
In this mode, which is selected when the `OUTPUT_ATTRS` environment
variable is set, the executable expects its input to be a list of
derivations and their corresponding Nix attributes; e.g.:

```
/nix/store/k60jbm15z02s6gf6g30cya9hq7s1pp1y-with-nix-eval-jobs.drv packages.x86_64-linux.with-nix-eval-jobs
/nix/store/j4ic0fr3q4jvg4nk88w1z63zdhwi027z-nix-buildkite-plugin-exe-nix-buildkite-0.2.0.0.drv packages.aarch64-darwin.nix-buildkite-plugin:exe:nix-buildkite
```

The plugin will then produce a Buildkite pipeline in which each
command is a `nix build` command (on Nix attributes), rather than a
`nix-store` command (on derivations).

We also add support for the `NIX_BUILD_OPTS` (analogous to
`NIX_STORE_OPTS`) and `OUTPUT_ATTRS_PREFIX` environment variables. The
latter prefixes each `nix build` command's attribute with the
specified string, in order to support flakes.

Signed-off-by: Drew Hess <src@drewhess.com>