feat: skip cached builds by default
We implement this, for now, anyway, by simply echoing a message to
`stdout` that the derivation/attribute has already been built. This
ensures that Buildkite reports the status of the build as "success".

(This can probably be further optimized, but we'll leave that for
subsequent work.)

We also provide a way to ignore the cache status, via the
`ignore-cache-status` plugin option.

Signed-off-by: Drew Hess <src@drewhess.com>